Split enums into different files
