+++ src/billing/types/enums/cardType.ts
@@ -1,0 +1,9 @@
+/**
+ * Possible card types
+ */
+export enum CardType {
+  VISA = 'Visa',
+  MASTERCARD = 'Mastercard',
+  MAESTRO = 'Maestro',
+  MIR = 'МИР'
+}
