+++ src/billing/types/enums/currency.ts
@@ -1,0 +1,7 @@
+/**
+ * Payment currency
+ */
+export enum Currency {
+  USD = 'USD',
+  RUB = 'RUB'
+}
