+++ src/billing/types/enums/interval.ts
@@ -1,0 +1,8 @@
+/**
+ * Reccurent payments interval
+ */
+export enum Interval {
+  WEEK = 'Week',
+  MONTH = 'Month',
+  DAY = 'Day'
+}
