+++ src/billing/types/enums/operationStatus.ts
@@ -1,0 +1,14 @@
+/**
+ * Payment status in case of successful completion
+ */
+export enum OperationStatus {
+  /**
+   * Status for one-step payments,
+   */
+  COMPLETED = 'Completed',
+
+  /**
+   * Status for two-step payments
+   */
+  AUTHORIZED = 'Authorized'
+}
